1.2.9
- HUD for v1 airships is done!  This will only show up if you are in a
v1 airship of any color.  There is a big % bar that lets you know how
much max fuel you have (up to a stack of 64), a small % bar that lets
you know how much current fuel you have (current fuel burn time
remaining %), a timer that gives you in second the max flight time with
all current fuel in the airship, and lights that indicate it is on!

Just have to copy it and make a v2 hud.