2.1.6
Huge update inc.

- Added 4 new items for the module system: Small Inventory Module, Large
Inventory Module, Fuel Efficiency Module, and the Infinite Fuel Module.

- Rewrote the V1 GUI structure.  Everything was changed to flow better
with the module system.  This includes the logic as well as the
images/layout.

- Fixed shift clicking items to and from the player inventory into: the
fuel slot on the default gui, the module slot in the module gui, and the
9 slot airship inventory/fuel slot in the small inv module gui.

-Fixed the fuel slot only allowing coal or charcoal in it.  Not more
confusion.  Also carried this over to the module slot, making it so only
modules can go in the slot.

-The HUD now gets its variables from the Core airship entity the player
is in.  This fixed a lot of the sync issues players were experiencing.
(No more having to open the gui to see the correct times/% after each
log in.)

- Fixed airships not registering being on in the client after reloging /
server restarting.  You no longer have to "take control"!

- Cleaned up and optimized a lot of the airship logic. It no longer does
a lot more work in the background than it needs to, thanks to my new
integrated boolean system overlayed on top of the new logic and
efficiently using my variables.

- Still a lot to do.  Must clean up the working code, add the Large Inv
container/gui/logic, test a lot, propogate all changes through the v2
and v3 airships, correct labels/names, etc.