Core changes for all versions of airship.
-Fixed a ticking entity client-side crash.  The damage calculations were
happening on both the server and the client side.  Now the calculations
happen only on the server and is then passed to the client.

Explosions no longer happen when breaking the airship back into item
form.  Changed the sound when the going into the item form so it makes
sense.

Reorganized the read / write priorities so read it first, write is
second.  Moved it to the top of the class file so it reads before doing
all of the calculations.