2.5.9
Upgrade the Eclipse workspace to the new recommended forge build 2099.

Added two new active modules : the Stealth Module and the Dash Module.

Updated the V1 Default GUI to detect if one of the new modules are
installed to turn on the green light.

Updated the V1 HUD to show the new modules when they are in the module
slot.

Implemented a cooldown timer that ticks down when "C" is pressed.  This
will get the tick amount from the module slot item if it is an active
module and use that to start the countdown.  This is saved to the
airship and another module cannot be activated while an airship is on
cooldown.  The cooldown calculations are correctly being passed to the
client.

The module slot will now recognize the two new modules.