2.6.2
Created a working cooldown system.  This is per airship and is tracked
by checking the id of the players riding entity and the airship itself.

Created a working Ability on system.  This also tracked by checking the
id of the players riding entity and the airship itself.

Tried to set the size and positioning of the V1 model to prevent it from
glitching into the ground.