3.0.3
Added a new fuel helper file to keep track of int values.

It looks the the gui's for the most part are working.  It appears to
have issues with being in creative mode.  It doesn't take the items int
value in the inv[0] fuel slot.  Will work around it.

HUD/GUI uses the getTotalPowered() int for calculations.

Adjusted the airship speed with the infinite fuel module to 60%, (up
from 50%).

-Changed the Airship balloon.  Used all leather!  No more hides.