3.1.0
-Fixed the calculations for the HUD Big Bar. I really overhaulled this
under the hood.  It was so wastful before keeping track of ticks
(sometimes 120,000+) when all I needed to do was a simple % of a max
stack of 64 items...  Woot!

-Fixed the calculations for the HUD Small Bar.

-Fixed the calculations for the HUD Timer.

-Added an new setup in the HUD and all GUIs for creative.  It will not
represent the infinite fuel module with the infinity symbols for the
timer and max green bars.  This does not over write the actual stored
burn time of the airship, only pauses it when a creative player is
piloting it.

Updated the GUI module screen to be grey when no module is in, and green
when a module is actively in, just because I could and it looks cooler
in my opinion!

-The airship now stores all variables used in the HUD and GUI that it
needs.