3.1.5
Updated the main, control, and module guidebooks.  They were kind out
very out dated...  This includes new information and updated pics.

All of the GUI's have been updated and all ariship types and variants
now have the ability for use the new fuel system.

When in creative, all % bars are filled in all GUIs and the timer is set
for infinite.

Again...  So much code cleanup.  Just keeping it real.