Full v4.7.2a Release + 1.11.2 v4.7.3
All changes have been applied to 1.11.2 and 1.10.X.

-----For 1.11.2 v4.7.3:
Restructured some of the main methods that take a while to redo when
back reving to 1.10.X.  This includes the container - transferStack and
the airship items - subitems.  This also compacted and mod a bit more.

Added @Override to various method to ensure the correct logic is being
used.