More fixes!~
- Fixed the airship GUI module light show up in the correct place.

- Adjusted the hitbox of all airships.  They are smaller, which helps
when shooting from an airship, but isn't 100% perfect.  It probably wont
be since it is how minecraft workd their hitboxes.  The same issue
exists with the boat.  It is better though.

- The glitching through the world issue is now 100% fixed.  I had to
rewrite everything i did and added specific methods to deal with it.  If
an airship is in "solid ground", the Y motion will override to 0.  It
does not affect player manuvering, only airship natural, unhindered
gravity.