The march!
- Added a new module: Module - Water Landing.  This module prevents
airships from sinking in water. If a player tried to manually go in the
water, as soon as the airship begins to smoke, the module overrides the
airship movement to put it back above water.
WIP - The module manual needs to be updated.

- The module system was revamped!  A lot of the core logic was cleaned
up and it is now easier to add future modules. Improved how modules
interact with airship logic. Clean up old unused logic.

- Added the Water Landing module display screen on all airships and the
icon will show up on the HUD.

- Cleaned up the Gui Module logic to more efficiently check if there is
a module in the slot.

- Added the Water Landing icon to the list that shows up on the HUD.

- Added the Water Landing icon to the display screen in all airships.

- Changed the Infinite Fuel modifier to slow an airship down by 50% of
the max speed (base speed + speed frame modifier).

- Removed the functionality of the old modules. The items are still in
the game but are Labeled as "OLD NAME". You can convert these old items
into the new ones for free by just putting them in the Airship
Workbench.