Real progress inside.
- Added a new Module: Max Altitude. This module removes the altitude
limitation of airships with no penalities. If the module is removed
while the ship is above the max altitude, it will rapidly call until it
is at the max.

- Added a new Module: Minor Efficiency. This module makes fuel last 25%
longer with no penalites.

- Added a new Module: Major Efficiency. This module makes fuel last 50%
longer with a -30% decrease in max speed.

NOTE: still need to update the crafting recipes and module guide.

-I rewrote my fuel system to accomidate the future energy system.

- Readjusted my fuel timers and % bars to work with my new system.

- Adjusted the max height calculation to be closer to the actual Y value
instead of being 1 under.

All airship max altitudes have been reduced by 10. The following are no
longer Unlimited: Diamond (230), Emerald (240), Prismarine (250), and
Purpur (260).