Combo release 4.30.29 - New update WIP
Release 4.30.29 -
-Version updated to 4.30.29 for 1.11.2/1.10.X.

New update WIP-
-The HUD Gui has been updated! The textures have been upgraded.  The %
bars are more accurate. The HUD now lists the current and max altitudes
(the Max Altitude Module will affect the HUD).  The closer you get to
the max, the colors of the numbers will change:

With in 15 blocks from the max, both the current and max will turn
yellow.
With in 5 blocks from the max, both the curretn and max will turn red.