Ebon Arts removal / Viesis' Gemstones conversion-update
- Reorganized all of the Creative tabs for future expansions.

- All classic Ebon Arts Gemstone Blocks (minus the cryptic) have been
converted over.

- All original and dense Gemstone Items have been converted over.

- Developed an enum system to spawn off block variants more efficiently.

- Updated the configuration system.  It will now save things more
efficiently and create a folder to put the file it.

- Created the Gemstone Cutter Workbench.  This item allows users to put
in items like Cobblestone and process them for a chance to cut gems out
of them.  You can choose what gem you want to find.  It is all working.
There are no longer Gemstone Tiers.  All gemstones are considered equal,
and when my weapon system comes out, they will be used a
modifiers/affects.
//TODO add % chances to find gems depending on the material to process.
Also try and fix the hopper issue where it will take out the fuel before
the gemstones... Also create the model for the cutter.

- Fixed/updated all core block logic.  I was sooooo young then I made
the first mod... <3

- Completely updated my packet handler system.

- Removed all ores (for now) and crops (forever!).

-Touched up a lot of the textures.  Still have to redo Carnelian and
Onyxius.