VG - Finishing touches!
- Creative tabs have been readjusted. There is now a General, Items, and
Regular Blocks tab.  All Ores/Workbenches will be in the General tab.

- Added Gem-Infused Ore. This ore drops a new item, Unidentified
Gemstone and also has a chance to drop random gemstones. Silk touch will
affect my ores, and will drop exp. There is a block for Overworld,
Nether, and Ender. WIP
//TODO Add to world generation.

- Started coming up with potion effects. WIP

- All rarity for gems have been readjusted.

- Fixed GUI workbench to respect on button boolean working on first
click.  Updates have been correctly synced.

- Fixed GUI Chance % to have a max of 500%.  Realigned the positioning
of the % based on the size.

- When looking at a workbench, name plate info will now be shown.  When
off, Gemstone Workbenches will list the "Gemstone" they are looking for
and that the "Power" is "Off".  When powered on, the power is show as
"On", the Gemstone they are looking for is listed, and a % bar ticks up
with changing colors when closer to completion. Regardless, if the
output slot is full, above the % will say "Full".

- Created a base block class that will pass on general settings to child
blocks. Woot for efficiency!

- The Gemstone Workbench now has a new model, looks amazing! All items
in the "fuel" slot show up the the model "glass" container. Blocks and
items now properly scale.  The Gemstone being looked for appears just
above the model on the "pedestal".

- Put the baseline Armor info in. WIP

- The Gemstone Workbench has been fixed to correctly save itemstack
info!

- Added a multi-proc chance to some items.  The % for some items/blocks
can be as high as 500%. Default Block % is 25%, Item is 5%. A higher %
chance mean that there is a chance to get an extra item for each 100%
above. I.E.- 500% would have a chance to drop 1-5 gems.

- Spend a lot of time on the gem proc library. A lot of items and blocks
have been defined. This is an ongoing labor of love.

- Added the ability to use .OBJ files for models.  Using none now, but
will eventually.

- Fixed a lot of localized names.  Blah.