ViesCraft all versions / Viesis' Gemstones - more changes!
ViesCraft
----------
For all:
- Fixed an issue where the crafting table would not drop items in the
last slot when breaking.

For 1.12:
- A lot of cleanup.
- Idea for a frame colorizer. WIP
- Rewrote the whole soudn system to make more sense.
- The airship appearance tab Done button will now exit back to the main
airship gui instead of just closing out.
- The airship appearance tab top label now states this is for "Frame
Appearance" with some extra textures drawn to make it easier to read.
Did this to plan for more customization in the future.
- Airship workbenches will now show the rendered item in the correct
slot just by looking at the model. I wanted to do this for a long time
now.  It will display and items put in, even from other mods.
- Updated the Airship Workbench texture to work with the new item
rendering model system.
- Fixed the Airship Workbench syncing issue.  It wasn't registering
items being in the slots until a player clicked on the workbench.

Viesis' Gemstones
--------------------
- Added a tool tip for the Gemstone cutter.