Another push for my 6.0.0 update.
- Rearranged my Display symbol enum to make more sense and allow future
expansion without having to mess things up for people in the future.
- Added more buttons to be able to be used.
- Adding holiday Display items! These will be special holiday themed
display symbols only available around the seasons. If you use one, it
will stay after the holiday's are over. If a player is in creative, they
will see all the holidays and can choose from them. So far I have :
- New Years
- Valentine's Day
- Easter
- 4th of July
- Halloween
- Thanksgiving
- Christmas
- Improved on the drawItemStack method used by all guis.
- The metaInfo in the guis will now correctly show the current airship's
choice instead of defaulting to a generic one each time the gui changes.
- Made a base class that extents to all gui. This has the special fonts,
item renderers and entity renderer logic to share.