Another huge one, INC!
- Reorganized my Ark folder to better make sense.
- Added Customization cost values for Core Models.
- Updated enums for visual textures of frames and balloons so that meta
0 reads "default", not "none". Just makes more sense.
- Recreated all of my buttons in the GuiVC. They are easier to use and
no longer have false labels.
- Airships not can freely change frames and balloons (the frame and
balloon part of this section) whil keeping their main textures.
- Disabled the multipass on the balloons. Will save for a future idea.
- A lot more gui customizations. The Frames and Balloons sections are
now fully done and are working with their costs.
- The Color guis of the frame and balloon now actively show you the
colors on change, not when you save. If you put any incorrect characters
in, it defaults to 0. If you go above the number cap, it defaults to
255.  This works a lot better now.
- The airship preview in the guis should now correctly apply those
independant changes for the sections they are in. For example, if I am
in Core -> Frame, I will see what my airship will look like with a new
model for the frame with my current textures, patterns, etc.
- Created a new GUI button that mimics the GuiButtonGeneralVC buttons
(This was my general button, now labeled GuiButtonGeneral1VC) to provide
more color options.